requirements.txt
pyTelegramBotAPI==4.14.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
APScheduler==3.10.4